Extraction of ContentQuery -> NodeInterface conversion in standalone service
needed for UI
