+++ Neos.ContentGraph.DoctrineDbalAdapter/Classes/Domain/Repository/NodeFactory.php
@@ -13,10 +13,22 @@ namespace Neos\ContentGraph\DoctrineDbalAdapter\Domain\Repository;
  */
 use Neos\ContentGraph\DoctrineDbalAdapter\Domain\Projection\Node;
 use Neos\ContentRepository\Domain as ContentRepository;
+use Neos\ContentRepository\Domain\Context\Parameters\ContextParameters;
+use Neos\ContentRepository\Domain\Model\NodeInterface;
 use Neos\ContentRepository\Domain\Projection\Content as ContentProjection;
+use Neos\ContentRepository\Domain\Projection\Content\ContentGraphInterface;
+use Neos\ContentRepository\Domain\Projection\Content\ContentSubgraphInterface;
+use Neos\ContentRepository\Domain\Projection\Workspace\WorkspaceFinder;
+use Neos\ContentRepository\Domain\ValueObject\NodeName;
 use Neos\ContentRepository\Domain\ValueObject\NodeTypeName;
 use Neos\ContentRepository\Utility;
 use Neos\Flow\Annotations as Flow;
+use Neos\Flow\Utility\Now;
+use Neos\Neos\Domain\Context\Content\ContentQuery;
+use Neos\Neos\Domain\Projection\Site\Site;
+use Neos\Neos\Domain\Projection\Site\SiteFinder;
+use Neos\Neos\Domain\Service\ContentContext;
+use Neos\Flow\Security\Context as SecurityContext;
 
 
 /**
@@ -29,9 +41,100 @@ final class NodeFactory
 
     /**
      * @Flow\Inject
+     * @var ContentGraphInterface
+     */
+    protected $contentGraph;
+
+    /**
+     * @Flow\Inject
+     * @var WorkspaceFinder
+     */
+    protected $workspaceFinder;
+
+
+    /**
+     * @Flow\Inject
+     * @var SiteFinder
+     */
+    protected $siteFinder;
+
+    /**
+     * @Flow\Inject(lazy=false)
+     * @var Now
+     */
+    protected $now;
+
+    /**
+     * @Flow\Inject
+     * @var SecurityContext
+     */
+    protected $securityContext;
+
+    /**
+     * @Flow\Inject
      * @var ContentRepository\Service\NodeTypeManager
      */
     protected $nodeTypeManager;
+
+    /**
+     * @param $contentQuery
+     * @return array
+     */
+    public function findNodeForContentQuery(ContentQuery $contentQuery): NodeInterface
+    {
+        $inBackend = !$contentQuery->getWorkspaceName()->isLive();
+        $workspace = $this->workspaceFinder->findOneByName($contentQuery->getWorkspaceName());
+        $subgraph = $this->contentGraph->getSubgraphByIdentifier(
+            $workspace->getCurrentContentStreamIdentifier(),
+            $contentQuery->getDimensionSpacePoint()
+        );
+
+        $contextParameters = $this->createContextParameters($inBackend);
+
+        $siteNode = $subgraph->findNodeByNodeAggregateIdentifier($contentQuery->getSiteIdentifier());
+        // TODO CACHE
+        $site = $this->siteFinder->findOneByNodeName(new NodeName($siteNode->getName()));
+
+        $contentContext = $this->createContentContext($contentQuery, $subgraph, $contextParameters, $site);
+
+        $node = $subgraph->findNodeByNodeAggregateIdentifier($contentQuery->getNodeAggregateIdentifier(), $contentContext);
+
+        return $node;
+    }
+
+    /**
+     * @param bool $inBackend
+     * @return ContextParameters
+     */
+    protected function createContextParameters(bool $inBackend): ContextParameters
+    {
+        return new ContextParameters($this->now, $this->securityContext->getRoles(), $inBackend, $inBackend, $inBackend);
+    }
+
+    /**
+     * @param ContentQuery $contentQuery
+     * @param ContentSubgraphInterface $subgraph
+     * @param ContextParameters $contextParameters
+     * @return ContentContext
+     */
+    protected function createContentContext(ContentQuery $contentQuery, ContentSubgraphInterface $subgraph, ContextParameters $contextParameters, Site $site): ContentContext
+    {
+        return new ContentContext(
+            (string)$contentQuery->getWorkspaceName(),
+            $contextParameters->getCurrentDateTime(),
+            $subgraph->getDimensionSpacePoint()->toLegacyDimensionArray(),
+            $subgraph->getDimensionSpacePoint()->getCoordinates(),
+            $contextParameters->isInvisibleContentShown(),
+            $contextParameters->isRemovedContentShown(),
+            $contextParameters->isInaccessibleContentShown(),
+            $site,
+            null,
+            $contentQuery->getRootNodeIdentifier(),
+            $subgraph,
+            $contextParameters
+        );
+    }
+
 
     /**
      * @param array $nodeRow Node Row from projection (neos_contentgraph_node table)
